# decrease allocations in residual evaluation