--- a/Project.toml
+++ b/Project.toml
@@ -16,7 +16,7 @@ SHA = "ea8e919c-243c-51af-8825-aaa63cd721ce"
 CodecBzip2 = "0.7"
 DataFrames = "1.2"
 JLD2 = "0.4"
-NLPModels = "0.17"
+NLPModels = "0.17, 0.18"
 julia = "^1.3.1"
 
 [extras]