This PR's objective is to decrease the amount of allocations when evaluating the residual. On problem-49-7776-pre/ladybug, I get

#### Before

```
julia> @benchmark residual!($nls, $(nls.meta.x0), $r)
BenchmarkTools.Trial: 109 samples with 1 evaluation.
 Range (min … max):  35.592 ms … 98.198 ms  ┊ GC (min … max):  0.00% … 7.13%
 Time  (median):     43.779 ms              ┊ GC (median):    13.09%
 Time  (mean ± σ):   46.145 ms ±  9.294 ms  ┊ GC (mean ± σ):  12.04% ± 3.36%

       ▃▃▃██▅▆
  ▅▅▁▁▇███████▇█▁▅▅▅▅▅▁▁▁▅▁▅▁▁▁▅▁▁▁▁▁▁▁▁▁▁▅▁▁▁▁▁▁▁▁▁▅▁▁▁▁▁▅▁▅ ▅
  35.6 ms      Histogram: log(frequency) by time      87.3 ms <

 Memory estimate: 46.31 MiB, allocs estimate: 1156139.
```

#### After

```
 julia> @benchmark residual!($nls, $(nls.meta.x0), $r)
BenchmarkTools.Trial: 129 samples with 1 evaluation.
 Range (min … max):  31.432 ms … 67.156 ms  ┊ GC (min … max):  0.00% … 0.00%
 Time  (median):     39.505 ms              ┊ GC (median):    16.44%
 Time  (mean ± σ):   38.756 ms ±  4.884 ms  ┊ GC (mean ± σ):  12.02% ± 7.62%

    ▂               ▅▆▄█▁
  ▅▅█▆▆▆▄▄▃▅▁▁▁▁▃▅▅▅█████▆▅▄▃▃▃▃▄▁▁▁▁▁▁▁▁▁▁▁▁▃▁▁▃▃▁▁▁▁▁▁▁▁▁▁▃ ▃
  31.4 ms         Histogram: frequency by time        56.2 ms <

 Memory estimate: 34.65 MiB, allocs estimate: 1028765.
 ```

There are still lots of allocations that don't seem necessary.

Any ideas?

cc @AntoninKns