@dpo 2022-02-06
```
julia> @benchmark residual!($nls, $(nls.meta.x0), $r)
BenchmarkTools.Trial: 143 samples with 1 evaluation.
 Range (min … max):  28.426 ms … 66.456 ms  ┊ GC (min … max):  0.00% … 16.05%
 Time  (median):     36.903 ms              ┊ GC (median):    19.57%
 Time  (mean ± σ):   35.035 ms ±  4.840 ms  ┊ GC (mean ± σ):  11.30% ±  9.94%

         ▄█▄▅▂                          ▂█ ▂▄▂▂                
  ▅▅▅▁▅▃▇█████▇▅▅▃▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▅▆▆▅▇██▇████▇▅▅▆▁▃▁▃▅▁▁▁▁▃ ▃
  28.4 ms         Histogram: frequency by time        42.5 ms <

 Memory estimate: 27.36 MiB, allocs estimate: 933236.
```