@dpo 2022-02-06
```
julia> @benchmark residual!($nls, $(nls.meta.x0), $r)
BenchmarkTools.Trial: 517 samples with 1 evaluation.
 Range (min … max):  7.942 ms … 24.299 ms  ┊ GC (min … max): 0.00% … 19.46%
 Time  (median):     9.010 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   9.676 ms ±  1.912 ms  ┊ GC (mean ± σ):  5.17% ±  9.69%

   ▁▁▅█▇▁▁ ▇▂                                                 
  ▅██████████▄▄▃▂▃▃▁▃▂▂▁▂▃▁▂▂▃▃▃▄▄▄▄▄▃▃▁▃▂▃▂▂▂▁▂▁▁▂▁▂▁▂▁▂▁▁▂ ▃
  7.94 ms        Histogram: frequency by time        16.3 ms <

 Memory estimate: 7.29 MiB, allocs estimate: 95531.
```