@dpo 2022-02-06
Something is going on with the way threading is used in here. It allocates tons of memory.