@dpo 2022-02-06
For comparison, here's the benchmark of the version with threading and 4 threads:

```
julia> @benchmark residual!($nls, $(nls.meta.x0), $r)
BenchmarkTools.Trial: 412 samples with 1 evaluation.
 Range (min … max):   7.387 ms … 50.271 ms  ┊ GC (min … max):  0.00% … 77.75%
 Time  (median):      8.254 ms              ┊ GC (median):     0.00%
 Time  (mean ± σ):   12.126 ms ±  9.268 ms  ┊ GC (mean ± σ):  29.92% ± 26.08%

  ▅█▅▁                                                 ▁
  ████▇▇▄█▇▄▁▁▄▁▄▁▁▄▁▄▄▁▁▁▄▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█▇███▅▅▅▆ ▆
  7.39 ms      Histogram: log(frequency) by time      37.4 ms <

 Memory estimate: 27.34 MiB, allocs estimate: 931745.
```