@dpo 2022-02-10
Success at last!
```
julia> @benchmark residual!($model, $(model.meta.x0), $r)
BenchmarkTools.Trial: 921 samples with 1 evaluation.
 Range (min … max):  4.857 ms …  10.066 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     5.321 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   5.413 ms ± 539.166 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

   ▁▆▅█▇▃▃▃▆▃▂                                                 
  ▆███████████▇▆▄▅▄▄▃▂▂▁▃▂▁▂▁▂▁▂▁▁▁▁▁▁▂▁▁▁▁▁▂▁▁▂▁▁▁▂▁▁▁▁▁▁▁▁▂ ▃
  4.86 ms         Histogram: frequency by time        8.77 ms <

 Memory estimate: 0 bytes, allocs estimate: 0.
```
Needs https://github.com/JuliaSmoothOptimizers/NLPModels.jl/pull/393 (otherwise, there remain 2 bytes of allocations!)