@dpo 2022-02-16
Non-allocating `jac_structure!` (requires NLPModels 0.18.2):
```
julia> @benchmark jac_structure!($model, $rows, $cols)
BenchmarkTools.Trial: 2150 samples with 1 evaluation.
 Range (min … max):  2.040 ms …   4.936 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     2.228 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   2.321 ms ± 313.718 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

     █▁                                                        
  ▅▆▅██▆▆▅▅▆▆▅▃▄▃▄▃▃▃▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▁▂▂▁▂▂▂▂▁▂▂▁▂▂▂▂▁▁▂▁▁▂▂▂▂ ▃
  2.04 ms         Histogram: frequency by time        4.04 ms <

 Memory estimate: 0 bytes, allocs estimate: 0.
```